Extracts `ImageDataStoreSpy` into a shared scope
